Finish basic frontend routing, add back button and minimal css with flexbox
