+++ frontend/my-app/src/App.css
@@ -7,5 +7,12 @@
 
 .card {
   padding: 2em;
+}
 
+.flexboxCalendar {
+  display: flex;
+  justify-content: center;
+  align-content: center;
+  align-items: center;
+  flex-direction: row;
 }
