+++ frontend/my-app/src/App.jsx
@@ -1,40 +1,13 @@
 import './App.css'
-import {useEffect, useState} from "react";
-import CalendarComp from "./Components/CalendarComp.jsx";
-
-const fetchAppointmentsForDay = (id) => {
-    return fetch(`/api/appointments/allforday/${id}`).then((res) => {
-        return res.json()
-    });
-}
+import {Link} from "react-router-dom";
 
 function App() {
-    const [bookedAppointments, setBookedAppointments] = useState([]);
-    const [selectedDay, setSelectedDay] = useState(new Date());
-
-    useEffect(() => {
-        fetchAppointmentsForDay("fetch")
-            .then((bookedAppointments) => {
-                console.log(bookedAppointments);
-                setBookedAppointments(bookedAppointments);
-            })
-    }, []);
-
-    const handleSelectDay = (selectedDay) => {
-        setSelectedDay(selectedDay);
-    }
-
     return (
         <>
             <div>Main Page</div>
-            <button className='reservationButton'>Reserve an appointment
-            </button>
-            <div>
-                <CalendarComp
-                     handleSelectedDay={handleSelectDay}
-                     bookedAppointments = {bookedAppointments}
-                />
-            </div>
+            <Link to="/calendar">
+                <button className='reservationButton' type="button">Reserve an appointment</button>
+            </Link>
         </>
     )
 }
