+++ frontend/my-app/src/Components/CalendarComp.jsx
@@ -1,33 +1,62 @@
-import {useState} from "react";
+import {useEffect, useState} from "react";
 import Calendar from 'react-calendar';
 import 'react-calendar/dist/Calendar.css';
 import Time from "./Time.jsx";
+import {Link} from "react-router-dom";
 
 function filterWeekends(date) {
     return date.getDay() === 0 || date.getDay() === 6;
 }
 
-const CalendarComp = ({handleSelectedDay, bookedAppointments}) => {
-
+const fetchAppointmentsForDay = (id) => {
+    return fetch(`/api/appointments/allforday/${id}`).then((res) => {
+        return res.json()
+    });
+}
+const CalendarComp = () => {
     const [date, setDate] = useState(new Date());
     const [showTime, setShowTime] = useState(false);
+    const [bookedAppointments, setBookedAppointments] = useState([]);
+    const [selectedDay, setSelectedDay] = useState(new Date());
+
+    useEffect(() => {
+        fetchAppointmentsForDay("fetch") // HARD CODED VALUE ONLY FOR TESTING PURPOSE
+            .then((bookedAppointments) => {
+                console.log(bookedAppointments);
+                setBookedAppointments(bookedAppointments);
+            })
+    }, []);
+
+    const handleSelectedDay = (selectedDay) => {
+        setSelectedDay(selectedDay);
+    }
 
     return (
         <div className="calendar">
-            <h1 className="header">IMF Blood Calendar</h1>
-            <div>
-                <Calendar
-                    onChange={setDate}
-                    value={date}
-                    onClickDay={() => {
-                        handleSelectedDay(date);
-                        setShowTime(true)}}
-                    tileDisabled={({ date}) => {
-                        const currentDate = new Date();
-                        currentDate.setHours(0, 0, 0, 0);
-                        return currentDate > date || filterWeekends(date);
-                    }}
-                />
+            <h2 className="header">IMF Blood Calendar</h2>
+            <div className="flexboxCalendar">
+                <div> {/*NAPTÁR*/}
+                    <Calendar
+                        onChange={setDate}
+                        value={date}
+                        onClickDay={() => {
+                            handleSelectedDay(date);
+                            setShowTime(true)
+                        }}
+                        tileDisabled={({date}) => {
+                            const currentDate = new Date();
+                            currentDate.setHours(0, 0, 0, 0);
+                            return currentDate > date || filterWeekends(date);
+                        }}
+                    />
+                </div>
+                <div>
+                    <Time
+                        showTime={showTime}
+                        date={date}
+                        bookedAppointments={bookedAppointments}
+                    />
+                </div>
             </div>
             {date.length > 0 ? (
                 <p>
@@ -41,13 +70,10 @@ const CalendarComp = ({handleSelectedDay, bookedAppointments}) => {
                 <p>
                     <span>Default selected date:</span>{date.toDateString()}
                 </p>
-            )
-            }
-            <Time
-                showTime={showTime}
-                date={date}
-                bookedAppointments = {bookedAppointments}
-            />
+            )}
+            <Link to="/">
+                <button className='backButton' type="button">BACK</button>
+            </Link>
         </div>
     );
 };
