+++ frontend/my-app/src/Components/Time.jsx
@@ -1,7 +1,6 @@
 import Times from "./Times.jsx";
 
 function Time({showTime, date, bookedAppointments}) {
-
     return (
         <div>
             {showTime ? <Times
